requirements.txt
yt-dlp
opencv-python
torch
numpy
scikit-learn
gradio
torchvision
Pillow
librosa
transformers
c2pa-python
google-genai
instaloader
requests
python-dotenv
PyWavelets
scipy
google-cloud-aiplatform
